GBIF2

Documentation for GBIF2.

GBIF2.GBIF2 — Module

GBIF2

The goals of GBIF2 is to follow the GBIF api as completely and correctly as possible.

Its main design features are:

Single results are Occurrence or Species objects with all GBIF fields available using object.fieldname, returning missing if not returned by a specific query.

Multiple results are returned as a Tables.jl compatible Table of Occurrence or Species rows. This Table can be converted to a DataFrame or writted directly to disk using CSV.jl and similar packages.

All GBIF enum keys are checked for correctness before querying so that only correct queries can be sent. Clear messages point to errors in queries.

A limit above 300 items at a time is allowed, unlike in the original API, by making multiple reuests and joining the results.

For even larger queries, download requests are handled with gbif.org account authentication.

A quick example

julia> using GBIF2, DataFrames

# Basic species match with `species_match`:
julia> sp = species_match("Lalage newtoni");

julia> sp.species
"Coracina newtoni"

julia> sp.synonym
true

julia> sp.vernacularName
missing

# Get a more detailed object with `species`:
julia> sp_detailed = species(sp);

julia> sp_detailed.vernacularName
"Reunion Cuckooshrike"

# Get the first 2000 occurrences for the species from 2000 to 2020, on reunion:
julia> oc = occurrence_search(sp;
           limit=2000, country=:RE, hasCoordinate=true, year=(2000,2020)
       ) |> DataFrame
2000×83 DataFrame
  Row │ decimalLongitude  decimalLatitude  year    month   day
      │ Float64?          Float64?         Int64?  Int64?  Int64?
──────┼────────────────────────────────────────────────────────────
    1 │          55.5085         -21.0192    2020       1      14
    2 │          55.4133         -20.928     2020       1      23
    3 │          55.4133         -20.928     2020       1      16
    4 │          55.5085         -21.0192    2020       1      14
    5 │          55.4123         -21.0184    2020       1      13
    6 │          55.4133         -20.928     2020       1      28
    7 │          55.4133         -20.928     2020       1      16
  ⋮   │        ⋮                 ⋮           ⋮       ⋮       ⋮
 1994 │          55.4133         -20.928     2017      10      29
 1995 │          55.4123         -21.0184    2017      10      25
 1996 │          55.4123         -21.0184    2017      10      25
 1997 │          55.4123         -21.0184    2017      10      17
 1998 │          55.4123         -21.0184    2017      10      25
 1999 │          55.4123         -21.0184    2017      10      25
 2000 │          55.4123         -21.0184    2017      10      25

GBIF2.GBIF2

GBIF2.Species

GBIF2.species

GBIF2.species_list

GBIF2.species_match

GBIF2.species_search

There are a number of ways to query the GBIF database for species, returning different numbers of results and amounts of data.

Species objects and queries correspond closely to the GBIF species api.

GBIF2.Species — Type

Species

Wrapper object for information returned by species, species_list , species_match or species_search queries. These often are species, but a more correctly taxa, as it may be e.g. "Aves" for all birds. We use Species for naming consistency with the GBIF API.

Species also serve as rows in Table, and are converted to rows in a DataFrame or CSV automatically by the Tables.jl interface.

Species properties are accessed with ., e.g. sp.kingdom. Note that these queries do not all return all properties, and not all records contain all properties in any case. Missing properties simply return missing.

The possible properties of a Species object are: (kingdom = Union{Missing, String}, phylum = Union{Missing, String}, class = Union{Missing, String}, order = Union{Missing, String}, family = Union{Missing, String}, genus = Union{Missing, String}, species = Union{Missing, String}, key = Union{Missing, Int64}, nubKey = Union{Missing, Int64}, nameKey = Union{Missing, Int64}, taxonID = Union{Missing, String}, sourceTaxonKey = Union{Missing, Int64}, kingdomKey = Union{Missing, Int64}, phylumKey = Union{Missing, Int64}, classKey = Union{Missing, Int64}, orderKey = Union{Missing, Int64}, familyKey = Union{Missing, Int64}, genusKey = Union{Missing, Int64}, speciesKey = Union{Missing, Int64}, datasetKey = Union{Missing, String}, constituentKey = Union{Missing, String}, scientificName = Union{Missing, String}, canonicalName = Union{Missing, String}, vernacularName = Union{Missing, String}, parentKey = Union{Missing, Int64}, parent = Union{Missing, String}, basionymKey = Union{Missing, Int64}, basionym = Union{Missing, String}, authorship = Union{Missing, String}, nameType = Union{Missing, String}, rank = Union{Missing, String}, origin = Union{Missing, String}, taxonomicStatus = Union{Missing, String}, nomenclaturalStatus = Union{Missing, Vector{String}}, remarks = Union{Missing, String}, publishedIn = Union{Missing, String}, numDescendants = Union{Missing, Int64}, lastCrawled = Union{Missing, String}, lastInterpreted = Union{Missing, String}, issues = Union{Missing, Vector{String}}, synonym = Union{Missing, Bool})

GBIF2.species — Function

species(key; kw...)
species(key, resulttype; kw...)

Query the GBIF species api, returning a single Species.

key: a species key, or Species object from another search that a key can be obtained from.

resulttype: set this so that instead of a Species, species will return an object in (:verbatim, :name, :parents, :children, :related, :synonyms, :combinations, :descriptions, :distributions, :media, :references, :speciesProfiles, :vernacularNames, :typeSpecimens). The return value will be a raw JSON3.Object, but itspropertynames` can be checked and used to access data.

Example

Here we find a species with species_search, and then obtain the complete record with species.

julia> using GBIF2
julia> tbl = species_search("Falco punctatus")
20-element GBIF2.Table{GBIF2.Species, JSON3.Array{JSON3.Object, Vector{UInt8}, SubArray{
UInt64, 1, Vector{UInt64}, Tuple{UnitRange{Int64}}, true}}}
┌──────────┬──────────┬───────────────┬───────────────┬────────────┬─────────┬──────────
│  kingdom │   phylum │         class │         order │     family │   genus │         ⋯
│  String? │  String? │       String? │       String? │    String? │ String? │         ⋯
├──────────┼──────────┼───────────────┼───────────────┼────────────┼─────────┼──────────
│ Animalia │  missing │          Aves │ Falconiformes │ Falconidae │ missing │ Falco p ⋯
│ Animalia │  missing │          Aves │       missing │ Falconidae │   Falco │ Falco p ⋯
│  missing │  missing │          Aves │ Falconiformes │ Falconidae │   Falco │ Falco p ⋯
│  missing │ Chordata │       missing │       missing │ Falconidae │   Falco │ Falco p ⋯
│ Animalia │ Chordata │          Aves │ Falconiformes │ Falconidae │   Falco │ Falco p ⋯
│ Animalia │ Chordata │          Aves │ Falconiformes │ Falconidae │   Falco │ Falco p ⋯
│  Metazoa │ Chordata │          Aves │ Falconiformes │ Falconidae │   Falco │ Falco p ⋯
│ Animalia │  missing │       missing │ Falconiformes │ Falconidae │ missing │ Falco p ⋯
│  Metazoa │ Chordata │          Aves │ Falconiformes │ Falconidae │   Falco │ Falco p ⋯
│    ⋮     │    ⋮     │       ⋮       │       ⋮       │     ⋮      │    ⋮    │         ⋱
└──────────┴──────────┴───────────────┴───────────────┴────────────┴─────────┴──────────
                                                          35 columns and 11 rows omitted

julia> species(tbl[6])
GBIF2.Species({
                   "key": 102091853,
                "nubKey": 2481005,
               "nameKey": 4400647,
               "taxonID": "175650",
               "kingdom": "Animalia",
                "phylum": "Chordata",
                 "order": "Falconiformes",
                "family": "Falconidae",
                 "genus": "Falco",
               "species": "Falco punctatus",
            "kingdomKey": 101683523,
             "phylumKey": 102017110,
              "classKey": 102085317,
              "orderKey": 102091762,
             "familyKey": 102091763,
              "genusKey": 102091765,
            "speciesKey": 102091853,
            "datasetKey": "9ca92552-f23a-41a8-a140-01abaa31c931",
             "parentKey": 102091765,
                "parent": "Falco",
        "scientificName": "Falco punctatus Temminck, 1821",
         "canonicalName": "Falco punctatus",
        "vernacularName": "Mauritius Kestrel",
            "authorship": "Temminck, 1821",
              "nameType": "SCIENTIFIC",
                  "rank": "SPECIES",
                "origin": "SOURCE",
       "taxonomicStatus": "ACCEPTED",
   "nomenclaturalStatus": [],
        "numDescendants": 0,
           "lastCrawled": "2022-10-10T18:15:33.989+00:00",
       "lastInterpreted": "2022-10-10T19:16:16.841+00:00",
                "issues": [
                            "SCIENTIFIC_NAME_ASSEMBLED"
                          ],
               "synonym": false,
                 "class": "Aves"
})

Keyword arguments

language: can be specified for a single argument or with second argument in (:parents, :children, :related, :synonyms).

datasetKey: can be specified, with a second argument :related.

GBIF2.species_list — Function

species_list(; kw...)
species_list(key; kw...)
species_list(key, resulttype; kw...)

Query the GBIF species_list api, returning a table of Species that exactly match your query.

Example

using GBIF2
species_list(; name="Lalage newtoni")

# output
8-element GBIF2.Table{GBIF2.Species, JSON3.Array{JSON3.Object, Vector{UInt8}, SubArray{UInt64, 1, Vector{UInt64}, Tuple{UnitRange{Int64}}, true}}}
┌──────────┬──────────┬───────────────┬───────────────┬───────────────┬──────────┬──────────────────┬───────────┬─────────┬──────────┬──────────────┬────────────────┬───────
│  kingdom │   phylum │         class │         order │        family │    genus │          species │       key │  nubKey │  nameKey │      taxonID │ sourceTaxonKey │ king ⋯
│  String? │  String? │       String? │       String? │       String? │  String? │          String? │    Int64? │  Int64? │   Int64? │      String? │         Int64? │      ⋯
├──────────┼──────────┼───────────────┼───────────────┼───────────────┼──────────┼──────────────────┼───────────┼─────────┼──────────┼──────────────┼────────────────┼───────
│ Animalia │ Chordata │          Aves │ Passeriformes │ Campephagidae │ Coracina │ Coracina newtoni │   8385394 │ missing │ 18882488 │ gbif:8385394 │      176651982 │      ⋯
│ Animalia │  missing │          Aves │       missing │ Campephagidae │   Lalage │   Lalage newtoni │ 100144670 │ 8385394 │  5976204 │        06014 │        missing │  128 ⋯
│ Animalia │  missing │          Aves │ Passeriformes │ Campephagidae │  missing │   Lalage newtoni │ 133165079 │ 8385394 │  5976204 │        18380 │        missing │  135 ⋯
│ Animalia │ Chordata │          Aves │ Passeriformes │ Campephagidae │   Lalage │   Lalage newtoni │ 161400685 │ 8385394 │ 18882488 │       895898 │        missing │  134 ⋯
│  missing │  missing │       missing │       missing │       missing │ Bossiaea │   Lalage newtoni │ 165585935 │ missing │ 18882488 │      6924877 │        missing │    m ⋯
│ Animalia │  missing │          Aves │ Passeriformes │ Campephagidae │   Lalage │   Lalage newtoni │ 165923305 │ 8385394 │ 18882488 │        19393 │        missing │  100 ⋯
│ Animalia │ Chordata │          Aves │ Passeriformes │ Campephagidae │   Lalage │   Lalage newtoni │ 168010293 │ 8385394 │  5976204 │       181376 │        missing │  167 ⋯
│ Animalia │ Chordata │ Passeriformes │          Aves │ Campephagidae │   Lalage │   Lalage newtoni │ 176651982 │ 8385394 │ 18882488 │     22706569 │        missing │  202 ⋯
└──────────┴──────────┴───────────────┴───────────────┴───────────────┴──────────┴──────────────────┴───────────┴─────────┴──────────┴──────────────┴────────────────┴───────

Keyword arguments

We use keywords exactly as in the GBIF api.

You can find keyword enum values with the [GBIF2.enum](@ref) function.

language: Language for vernacular names, given as an ISO 639-1 two-letter code from our

datasetKey: Filters by the checklist dataset key (a uuid)

sourceId: Filters by the source identifier

name: Name of the species

offset: Offset to start results from

limit: The maximum number of results to return. This can't be greater than 300, any value greater is set to 300.

GBIF2.species_match — Function

species_match(; kw...)

Query the GBIF species/match api, returning the single closest Species using fuzzy search.

The results are not particularly detailed, this can be improved by calling species(res) on the result of species_match to query for the full dataset.

Example

using GBIF2
sp = species_match("Lalage newtoni")

# output
GBIF2.Species({
           "usageKey": 8385394,
   "acceptedUsageKey": 2486791,
     "scientificName": "Lalage newtoni (Pollen, 1866)",
      "canonicalName": "Lalage newtoni",
               "rank": "SPECIES",
             "status": "SYNONYM",
         "confidence": 98,
          "matchType": "EXACT",
            "kingdom": "Animalia",
             "phylum": "Chordata",
              "order": "Passeriformes",
             "family": "Campephagidae",
              "genus": "Coracina",
            "species": "Coracina newtoni",
         "kingdomKey": 1,
          "phylumKey": 44,
           "classKey": 212,
           "orderKey": 729,
          "familyKey": 9284,
           "genusKey": 2482359,
         "speciesKey": 2486791,
            "synonym": true,
              "class": "Aves"
})

Keywords

We use keywords exactly as in the GBIF api.

You can find keyword enum values with the [GBIF2.enum](@ref) function.

rank: Filters by taxonomic rank as given in our Rank enum

name: Name of the species

strict: If true it (fuzzy) matches only the given name, but never a taxon in the upper classification

verbose: If true it shows alternative matches which were considered but then rejected

kingdom: Optional kingdom classification accepting a canonical name.

phylum: Optional phylum classification accepting a canonical name.

class: Optional class classification accepting a canonical name.

order: Optional order classification accepting a canonical name.

family: Optional family classification accepting a canonical name.

genus: Optional genus classification accepting a canonical name.

GBIF2.species_search — Function

species_search([q]; kw...)

Query the GBIF species/search api, returning many results in a GBIF2.Table.

Example

using GBIF2
sp = species_search("Psittacula eques")

# output
20-element GBIF2.Table{GBIF2.Species, JSON3.Array{JSON3.Object, Vector{UInt8}, SubArray{UInt64, 1,
Vector{UInt64}, Tuple{UnitRange{Int64}}, true}}}
┌──────────┬──────────┬────────────────┬────────────────┬───────────────┬──────────────┬───────────
│  kingdom │   phylum │          class │          order │        family │        genus │          ⋯
│  String? │  String? │        String? │        String? │       String? │      String? │          ⋯
├──────────┼──────────┼────────────────┼────────────────┼───────────────┼──────────────┼───────────
│ Animalia │ Chordata │           Aves │ Psittaciformes │ Psittaculidae │   Psittacula │   Psitta ⋯
│ Animalia │  missing │           Aves │ Psittaciformes │ Psittaculidae │      missing │          ⋯
│  Metazoa │ Chordata │           Aves │ Psittaciformes │   Psittacidae │   Psittacula │   Psitta ⋯
│ Animalia │  missing │           Aves │ Psittaciformes │ Psittaculidae │      missing │   Psitta ⋯
│ Animalia │  missing │           Aves │ Psittaciformes │ Psittaculidae │      missing │          ⋯
│ Animalia │  missing │        missing │ Psittaciformes │ Psittaculidae │      missing │   Psitta ⋯
│  Metazoa │ Chordata │           Aves │ Psittaciformes │   Psittacidae │   Psittacula │   Psitta ⋯
│  missing │  missing │        missing │        missing │       missing │   Psittacula │   Psitta ⋯
│ Animalia │ Chordata │           Aves │ Psittaciformes │ Psittaculidae │   Psittacula │   Psitta ⋯
│ Animalia │  missing │           Aves │        missing │   Psittacidae │   Psittacula │   Psitta ⋯
│  Metazoa │ Chordata │           Aves │ Psittaciformes │   Psittacidae │   Psittacula │    Psitt ⋯
│ Animalia │ Chordata │           Aves │ Psittaciformes │ Psittaculidae │   Psittacula │   Psitta ⋯
│ ANIMALIA │ CHORDATA │ PSITTACIFORMES │           AVES │   PSITTACIDAE │ Alexandrinus │ Alexandr ⋯
│  Metazoa │ Chordata │           Aves │ Psittaciformes │   Psittacidae │   Psittacula │    Psitt ⋯
│ Animalia │ Chordata │           Aves │ Psittaciformes │   Psittacidae │   Psittacula │   Psitta ⋯
│ Animalia │  missing │           Aves │ Psittaciformes │ Psittaculidae │   Psittacula │   Psitta ⋯
│    ⋮     │    ⋮     │       ⋮        │       ⋮        │       ⋮       │      ⋮       │          ⋱
└──────────┴──────────┴────────────────┴────────────────┴───────────────┴──────────────┴───────────
                                                                      35 columns and 4 rows omitted

Keyword arguments

We use keywords exactly as in the GBIF api.

class: Optional class classification accepting a canonical name.

datasetKey: Filters by the checklist dataset key (a uuid)

facet: A list of facet names used to retrieve the 100 most frequent values for a field.

Allowed facets are :datasetKey, :higherTaxonKey, :rank, :status, :nomenclaturalStatus, isExtinct, :habitat, :threat and :nameType.

facetMincount: Used in combination with the facet parameter. Set facetMincount=N to exclude facets

with a count less than N, e.g. facet=type, limit=>0, facetMincount=>10000 only shows the type value OCCURRENCE because :CHECKLIST and :METADATA have counts less than 10000.

facetMultiselect: Used in combination with the facet parameter. Set facetMultiselect=true to still

return counts for values that are not currently filtered, e.g. facet=type, limit=>0, type=>CHECKLIST, facetMultiselect=>true still shows type values OCCURRENCE and METADATA even though type is being filtered by type=:CHECKLIST

family: Optional family classification accepting a canonical name.

genus: Optional genus classification accepting a canonical name.

habitat: Filters by the habitat. Currently only 3 major biomes are accepted in our Habitat enum

highertaxonKey: Filters by any of the higher Linnean rank keys. Note this is within the respective",

checklist and not searching nub keys across all checklists.

hl: Set hl=true to highlight terms matching the query when in fulltext search fields.",

The highlight will be an emphasis tag of class 'gbifH1' e.g. q="plant", hl=>true.", Fulltext search fields include title, keyword, country, publishing country, publishing organization title,", hosting organization title, and description. One additional full text field is searched which includes", information from metadata documents, but the text of this field is not returned in the response.

isExtinct: Filters by extinction status (a boolean, e.g. isExtinct=>true)

issue: A specific indexing issue as defined in our NameUsageIssue enum

kingdom: Optional kingdom classification accepting a canonical name.

language: Language for vernacular names, given as an ISO 639-1 two-letter code from our

nameType: Filters by the name type as given in our NameType enum

nomenclaturalStatus: Not yet implemented, but will eventually allow for filtering by a nomenclatural status enum

order: Optional order classification accepting a canonical name.

phylum: Optional phylum classification accepting a canonical name.

q: Simple full text search parameter. The value for this parameter can be a simple word or a phrase.

Wildcards are not supported

rank: Filters by taxonomic rank as given in our Rank enum

sourceId: Filters by the source identifier

status: Filters by the taxonomic status as given in our TaxonomicStatus enum

strict: If true it (fuzzy) matches only the given name, but never a taxon in the upper classification

threat: Filters by the taxonomic threat status as given in our ThreatStatus enum

verbose: If true it shows alternative matches which were considered but then rejected

offset: Offset to start results from

limit: The maximum number of results to return. This can't be greater than 300, any value greater is set to 300.

Occurrence objects and queries correspond closely to the GBIF occurrence api.

Occurrence
occurrence
occurrence_search
occurrence_request
occurrence_download
occurrence_count
occurrence_inventory